Removed legacy loader calls from ParseIsotopeLabelTests and moved all tests to appropriate classes.
Files checked in: DataRepo/tests/management/commands/test_load_peak_annotations.py DataRepo/tests/test_models.py DataRepo/tests/utils/test_infusate_parser.py
